+++ src/js/modules/locator.js
@@ -1,6 +1,7 @@
 /**
  * Yuzora - Service Locator Module
  */
+"use strict";
 
 class Locator {
     /**